+++ src/main/java/edu/epam/bookie/service/BetService.java
@@ -6,11 +6,39 @@ import edu.epam.bookie.model.sport.Bet;
 import java.util.List;
 
 public interface BetService {
-    List<Bet> selectBetsByMatchId(long id) throws ServiceException;
+    /**
+     * Select all bets for this match
+     *
+     * @param id match id
+     * @return list of bets
+     * @throws ServiceException exception
+     */
+    List<Bet> selectBetsByMatchId(int id) throws ServiceException;
 
-    List<Bet> selectBetsByUserId(long id) throws ServiceException;
+    /**
+     * Select all bets of user
+     *
+     * @param id user id
+     * @return list of bets
+     * @throws ServiceException exception
+     */
+    List<Bet> selectBetsByUserId(int id) throws ServiceException;
 
+    /**
+     * Pay for bet and change bet status to 'WON"
+     *
+     * @param bet bet
+     * @return result
+     * @throws ServiceException exception
+     */
     boolean payBets(Bet bet) throws ServiceException;
 
+    /**
+     * Change bet status to 'LOST"
+     *
+     * @param bet bet
+     * @return result
+     * @throws ServiceException exception
+     */
     boolean betLost(Bet bet) throws ServiceException;
 }
